Create initial exchange provider
