+++ src/app/app.module.ts
@@ -8,6 +8,7 @@ import { MyApp } from './app.component';
 import { HomePage } from '../pages/home/home';
 import { AddEditTransactionPage } from '../pages/add-edit-transaction/add-edit-transaction';
 
+import { ExchangeProvider } from '../providers/exchange/exchange';
 import { TransactionProvider } from '../providers/transaction/transaction';
 
 @NgModule({
@@ -30,6 +31,7 @@ import { TransactionProvider } from '../providers/transaction/transaction';
     StatusBar,
     SplashScreen,
     {provide: ErrorHandler, useClass: IonicErrorHandler},
+    ExchangeProvider,
     TransactionProvider
   ]
 })
