+++ src/pages/add-edit-transaction/add-edit-transaction.ts
@@ -3,6 +3,7 @@ import { IonicPage, NavController, NavParams } from 'ionic-angular';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { TransactionProvider } from '../../providers/transaction/transaction';
 import { CryptoType } from '../../enums/crypto-type';
+import { ExchangeProvider } from '../../providers/exchange/exchange';
 
 /**
  * Create or edit a crypto currency transaction.
@@ -31,7 +32,8 @@ export class AddEditTransactionPage {
     complete: false
   };
 
-  constructor(public navCtrl: NavController, public navParams: NavParams, public formBuilder: FormBuilder, public transactionProvider: TransactionProvider) {
+  constructor(public navCtrl: NavController, public navParams: NavParams, public formBuilder: FormBuilder,
+              private exchangeProvider: ExchangeProvider, public transactionProvider: TransactionProvider) {
     this.transactionForm = formBuilder.group({
       exchange: ['', Validators.required],
       cryptoType: ['', Validators.required],
@@ -39,36 +41,17 @@ export class AddEditTransactionPage {
       cryptoQuantity: ['', Validators.required]
     });
 
-    this.initializeExchanges();
+    this.exchanges = exchangeProvider.getExchanges();
 
     if (this.navParams.get('transaction')) {
       this.transaction = this.navParams.get('transaction');
     } else {
-      this.transaction.exchange = this.getExchangeByName('Binance').name;
+      this.transaction.exchange = this.exchangeProvider.getExchangeByName('Binance').name;
       this.transaction.cryptoType = CryptoType.BTC;
     }
   }
 
   ionViewDidLoad() {}
-
-  /**
-   * Initialize the available exchanges.
-   */
-  initializeExchanges() {
-    this.exchanges = [];
-
-    let binanceExchange: Exchange = {
-      name: 'Binance',
-      transactionFeePercentage: 0.001
-    };
-    this.exchanges.push(binanceExchange);
-
-    let gdaxExchange: Exchange = {
-      name: 'GDAX',
-      transactionFeePercentage: 0
-    };
-    this.exchanges.push(gdaxExchange);
-  }
 
   /**
    * Calculate the transaction cost and suggested sale prices.
@@ -86,24 +69,12 @@ export class AddEditTransactionPage {
   }
 
   /**
-   * Retrieve an exchange and it's information by the exchanges name.
-   *
-   * @param exchangeName The name of the exchange.
-   * @returns {Exchange | undefined} The desired exchange and it's information.
-   */
-  getExchangeByName(exchangeName) {
-    return this.exchanges.find(function (exchange) {
-      return exchange.name === exchangeName;
-    });
-  }
-
-  /**
    * Calculate the cost of the transaction in the currency the user is buying in.
    *
    * @returns {number} The cost of the users purchase.
    */
   calculatePurchaseCost() {
-    return (this.transaction.cryptoQuantity * this.transaction.currentCryptoPrice) * (1 + (this.getExchangeByName(this.transaction.exchange).transactionFeePercentage * 2));
+    return (this.transaction.cryptoQuantity * this.transaction.currentCryptoPrice) * (1 + (this.exchangeProvider.getExchangeByName(this.transaction.exchange).transactionFeePercentage * 2));
   }
 
   /**
@@ -112,7 +83,7 @@ export class AddEditTransactionPage {
    * @returns {number} The break even price for the current transaction.
    */
   calculateBreakEvenPrice() {
-    return this.transaction.currentCryptoPrice * (1 + (this.getExchangeByName(this.transaction.exchange).transactionFeePercentage * 2));
+    return this.transaction.currentCryptoPrice * (1 + (this.exchangeProvider.getExchangeByName(this.transaction.exchange).transactionFeePercentage * 2));
   }
 
   /**
